+++ packages/pluggable-widgets-tools/configs/rollup.config.mjs
@@ -1,5 +1,3 @@
-/* eslint-disable @typescript-eslint/explicit-function-return-type */
-
 import { existsSync } from "node:fs";
 import { join } from "node:path";
 import { fileURLToPath } from "node:url";
@@ -122,7 +120,7 @@ export default async args => {
                 postCssPlugin(outputFormat, production),
                 alias({
                     entries: {
-                        "react-hot-loader/root": fileURLToPath(new URL("hot", import.meta.url)),
+                        "react-hot-loader/root": fileURLToPath(new URL("hot", import.meta.url))
                     }
                 }),
                 ...getCommonPlugins({
@@ -199,7 +197,7 @@ export default async args => {
                             setTimeout(() => process.exit(0));
                         }
                     },
-                    name: 'force-close'
+                    name: "force-close"
                 }
             ],
             onwarn: onwarn(args)
@@ -208,10 +206,11 @@ export default async args => {
 
     const customConfigPathJS = join(sourcePath, "rollup.config.js");
     const customConfigPathESM = join(sourcePath, "rollup.config.mjs");
-    const existingConfigPath =
-        existsSync(customConfigPathJS) ? customConfigPathJS
-            : existsSync(customConfigPathESM) ? customConfigPathESM
-                : null;
+    const existingConfigPath = existsSync(customConfigPathJS)
+        ? customConfigPathJS
+        : existsSync(customConfigPathESM)
+          ? customConfigPathESM
+          : null;
     if (existingConfigPath != null) {
         const customConfig = await loadConfigFile(existingConfigPath, { ...args, configDefaultConfig: result });
         customConfig.warnings.flush();
@@ -225,12 +224,12 @@ export default async args => {
             nodeResolve({ preferBuiltins: false, mainFields: ["module", "browser", "main"] }),
             isTypescript
                 ? typescript({
-                    noEmitOnError: !args.watch,
-                    sourceMap: config.sourceMaps,
-                    inlineSources: config.sourceMaps,
-                    target: "es2022", // we transpile the result with babel anyway, see below
-                    exclude: ["**/__tests__/**/*"]
-                })
+                      noEmitOnError: !args.watch,
+                      sourceMap: config.sourceMaps,
+                      inlineSources: config.sourceMaps,
+                      target: "es2022", // we transpile the result with babel anyway, see below
+                      exclude: ["**/__tests__/**/*"]
+                  })
                 : null,
             // Babel can transpile source JS and resulting JS, hence are input/output plugins. The good
             // practice is to do the most of conversions on resulting code, since then we ensure that
@@ -267,29 +266,29 @@ export default async args => {
             }),
             config.transpile
                 ? getBabelOutputPlugin({
-                    sourceMaps: config.sourceMaps,
-                    babelrc: false,
-                    compact: false,
-                    ...(config.babelConfig || {})
-                })
+                      sourceMaps: config.sourceMaps,
+                      babelrc: false,
+                      compact: false,
+                      ...(config.babelConfig || {})
+                  })
                 : null,
             image(),
             production ? terser() : null,
             config.licenses
                 ? license({
-                    thirdParty: {
-                        includePrivate: true,
-                        output: [
-                            {
-                                file: join(outDir, "dependencies.txt")
-                            },
-                            {
-                                file: join(outDir, "dependencies.json"),
-                                template: licenseCustomTemplate
-                            }
-                        ]
-                    }
-                })
+                      thirdParty: {
+                          includePrivate: true,
+                          output: [
+                              {
+                                  file: join(outDir, "dependencies.txt")
+                              },
+                              {
+                                  file: join(outDir, "dependencies.json"),
+                                  template: licenseCustomTemplate
+                              }
+                          ]
+                      }
+                  })
                 : null,
             // We need to create .mpk and copy results to test project after bundling is finished.
             // In case of a regular build is it is on `writeBundle` of the last config we define