+++ packages/pluggable-widgets-tools/src/typings-generator/generateClientTypes.ts
@@ -70,12 +70,15 @@ function generateClientTypeBody(
             }
             return true;
         })
-        .map(prop => `    ${prop.$.key}${isOptionalProp(prop, resolveProp) ? "?" : ""}: ${toClientPropType(
-            prop,
-            isNative,
-            generatedTypes,
-            resolveProp
-        )};`)
+        .map(
+            prop =>
+                `    ${prop.$.key}${isOptionalProp(prop, resolveProp) ? "?" : ""}: ${toClientPropType(
+                    prop,
+                    isNative,
+                    generatedTypes,
+                    resolveProp
+                )};`
+        )
         .join("\n");
 }
 
@@ -101,7 +104,8 @@ export function hasOptionalDataSource(prop: Property, resolveProp: (key: string)
 }
 
 function toActionVariablesOutputType(actionVariables?: ActionVariableTypes[]) {
-    const types = actionVariables?.flatMap(av => av.actionVariable)
+    const types = actionVariables
+        ?.flatMap(av => av.actionVariable)
         .map(avt => `${avt.$.key}: ${toOption(toAttributeClientType(avt.$.type))}`)
         .join("; ");
 
@@ -149,15 +153,16 @@ function toClientPropType(
 
             if (prop.$.isMetaData === "true") {
                 if (!linkedToDataSource) {
-                    throw new Error(`[XML] Attribute property can only have isMetaData="true" when linked to a datasource`);
+                    throw new Error(
+                        `[XML] Attribute property can only have isMetaData="true" when linked to a datasource`
+                    );
                 }
                 return `AttributeMetaData<${unionType}>`;
             }
 
             if (!prop.associationTypes?.length) {
                 return toAttributeOutputType("Reference", linkedToDataSource, unionType);
-            }
-            else {
+            } else {
                 const reftypes = prop.associationTypes
                     .flatMap(ats => ats.associationType)
                     .map(at => toAttributeOutputType(at.$.name, linkedToDataSource, unionType));
@@ -172,7 +177,9 @@ function toClientPropType(
             const linkedToDataSource = !!prop.$.dataSource;
             if (prop.$.isMetaData === "true") {
                 if (!linkedToDataSource) {
-                    throw new Error(`[XML] Association property can only have isMetaData="true" when linked to a datasource`);
+                    throw new Error(
+                        `[XML] Association property can only have isMetaData="true" when linked to a datasource`
+                    );
                 }
                 return "AssociationMetaData";
             }
@@ -319,9 +326,13 @@ export function toAssociationOutputType(xmlType: string, linkedToDataSource: boo
 export function toAttributeOutputType(xmlType: string, linkedToDataSource: boolean, unionAttributeType: string) {
     switch (xmlType) {
         case "Reference":
-            return linkedToDataSource ?  `ListAttributeValue<${unionAttributeType}>` : `EditableValue<${unionAttributeType}>`;
+            return linkedToDataSource
+                ? `ListAttributeValue<${unionAttributeType}>`
+                : `EditableValue<${unionAttributeType}>`;
         case "ReferenceSet":
-            return linkedToDataSource ?  `ListAttributeListValue<${unionAttributeType}>` : `EditableListValue<${unionAttributeType}>`;
+            return linkedToDataSource
+                ? `ListAttributeListValue<${unionAttributeType}>`
+                : `EditableListValue<${unionAttributeType}>`;
         default:
             return "any";
     }