+++ packages/pluggable-widgets-tools/test-config/__mocks__/hot.js
@@ -1,3 +1,3 @@
 jest.mock("react-hot-loader/root", () => ({
     hot: jest.fn(widget => widget)
-}))
+}));