+++ packages/pluggable-widgets-tools/test-config/jest-svg-transformer.js
@@ -1,15 +1,15 @@
-const path = require('path');
+const path = require("path");
 
 module.exports = {
     process(src, filePath) {
-        if (path.extname(filePath) !== '.svg') {
+        if (path.extname(filePath) !== ".svg") {
             return src;
         }
 
-        const name = `svg-${path.basename(filePath, '.svg')}`
+        const name = `svg-${path.basename(filePath, ".svg")}`
             .split(/\W+/)
-            .map((x) => `${x.charAt(0).toUpperCase()}${x.slice(1)}`)
-            .join('');
+            .map(x => `${x.charAt(0).toUpperCase()}${x.slice(1)}`)
+            .join("");
 
         return {
             code: `
@@ -23,5 +23,5 @@ function ${name}(props) {
 module.exports = ${name};
 `
         };
-    },
+    }
 };