+++ packages/pluggable-widgets-tools/utils/migration.js
@@ -1,10 +1,12 @@
 const { join } = require("path");
-const { createInterface } = require("readline");
 const { readJson, writeJson } = require("fs-extra");
 const { execSync } = require("child_process");
-const { red, green, yellow } = require("ansi-colors");
-const { copyFileSync, existsSync, mkdirSync, readFile, promises } = require("fs");
+const { red, green, yellow, whiteBright, bold } = require("ansi-colors");
+const { copyFileSync, existsSync, mkdirSync, promises } = require("fs");
 const { parseStringPromise } = require("xml2js");
+const { rm } = require("fs/promises");
+const { auditPluggableWidgetsTools } = require("../dist/commands/audit");
+const { confirm } = require("../dist/cli/confirm");
 let requirePatch = false;
 
 const CheckType = {
@@ -63,16 +65,6 @@ function extractVersions(version) {
     return version.replace(/^\D+/, "").split(".").map(Number);
 }
 
-async function question(question) {
-    const rl = createInterface({ input: process.stdin, output: process.stdout });
-    return new Promise(resolve =>
-        rl.question(yellow(question), answer => {
-            rl.close();
-            resolve(!answer ? "y" : answer.toLowerCase());
-        })
-    );
-}
-
 function getOutdatedDependencies(packageDependencies, listOfNewDependencies = dependencies) {
     return listOfNewDependencies
         .filter(dep => !!packageDependencies[dep.name])
@@ -100,7 +92,7 @@ function replaceOldDependencies(listOfOutdatedDependencies, packageJson, key) {
             } else {
                 packageJson[key][dep.name] = dep.newVersion;
 
-                if (!!dep.patch) {
+                if (dep.patch) {
                     const dir = join(process.cwd(), "patches");
                     if (!existsSync(dir)) {
                         mkdirSync(dir);
@@ -128,7 +120,7 @@ async function addExtraDependencies(packageJson, key) {
 
 async function getExtraDependencies(packageJson, key) {
     const sourceDir = process.cwd();
-    const rawPackageXML = await promises.readFile(join(sourceDir, "src/package.xml"), 'utf-8');
+    const rawPackageXML = await promises.readFile(join(sourceDir, "src/package.xml"), "utf-8");
     if (!rawPackageXML) {
         throw new Error("package.xml file was not found, please check your src folder");
     }
@@ -143,17 +135,26 @@ async function getExtraDependencies(packageJson, key) {
         .reduce((a, e) => a.concat(e), [])
         .filter(wfXml => wfXml.$.path);
     if (widgetDefinitionXMLPaths.length === 0)
-        throw new Error("Path to the widget definition XML file(s) could not be found, please check your package.xml file")
+        throw new Error(
+            "Path to the widget definition XML file(s) could not be found, please check your package.xml file"
+        );
 
     const parsedWidgetDefinitionXMLs = [];
     for (const widgetDefinitionXMLPath of widgetDefinitionXMLPaths) {
-        const rawWidgetDefinitionXML = await promises.readFile(join(sourceDir, 'src/', widgetDefinitionXMLPath.$.path), 'utf-8');
+        const rawWidgetDefinitionXML = await promises.readFile(
+            join(sourceDir, "src/", widgetDefinitionXMLPath.$.path),
+            "utf-8"
+        );
         if (!rawWidgetDefinitionXML) {
-            throw new Error(`Widget definition XML file (with path ${widgetDefinitionXMLPath}) could not be found, please check your src folder`)
+            throw new Error(
+                `Widget definition XML file (with path ${widgetDefinitionXMLPath}) could not be found, please check your src folder`
+            );
         }
         const parsedWidgetDefinitionXML = await parseStringPromise(rawWidgetDefinitionXML);
         if (!parsedWidgetDefinitionXML) {
-            throw new Error(`Widget definition XML file (with path ${widgetDefinitionXMLPath}) is empty, please check your src folder`)
+            throw new Error(
+                `Widget definition XML file (with path ${widgetDefinitionXMLPath}) is empty, please check your src folder`
+            );
         }
         parsedWidgetDefinitionXMLs.push(parsedWidgetDefinitionXML);
     }
@@ -169,7 +170,9 @@ async function getExtraDependencies(packageJson, key) {
 
     let extraDependencies = resolutionsOverrides.filter(ov => !packageJson[key] || !packageJson[key][ov.name]);
     if (!supportedPlatforms.includes("Native"))
-        extraDependencies = extraDependencies.filter(d => d.name !== "react-native" && d.name !== "@types/react-native")
+        extraDependencies = extraDependencies.filter(
+            d => d.name !== "react-native" && d.name !== "@types/react-native"
+        );
     if (!supportedPlatforms.includes("Web"))
         extraDependencies = extraDependencies.filter(d => d.name !== "react-dom" && d.name !== "@types/react-dom");
 
@@ -177,9 +180,14 @@ async function getExtraDependencies(packageJson, key) {
 }
 
 async function checkMigration() {
+    const paths = {
+        packageJson: join(process.cwd(), "package.json"),
+        packageLock: join(process.cwd(), "package-lock.json"),
+        node_modules: join(process.cwd(), "node_modules")
+    };
+
     console.log("Checking if dependencies should be migrated...");
-    const packageJsonPath = join(process.cwd(), "package.json");
-    const packageJson = await readJson(packageJsonPath);
+    const packageJson = await readJson(paths.packageJson);
     const args = process.argv;
     if (!args.includes("--skip-migration") && process.env.CI !== "true") {
         const outdatedDependencies = getOutdatedDependencies(packageJson.dependencies || {});
@@ -192,11 +200,15 @@ async function checkMigration() {
             outdatedOverrides.length > 0 ||
             outdatedResolutions.length > 0
         ) {
-            const answer = await question(
-                "Your widget contains outdated dependencies that will not work with this version of Pluggable Widgets Tools, do you want to upgrade it automatically? " +
-                "Note that this operation will delete your node_modules folder and package-lock.json files and re-create them. [Y/n]: "
+            console.log(
+                yellow(
+                    "Your widget contains outdated dependencies that will not work with this version of Pluggable Widgets Tools.\n" +
+                        (existsSync(paths.packageLock)
+                            ? "Note: To ensure a clean upgrade, a force-install is performed. Removing the lockfile and node_modules before installation."
+                            : "Note: To ensure a clean upgrade, a force-install is recommended. Removing the lockfile and node_modules before installation.")
+                )
             );
-            if (answer === "y") {
+            if (await confirm("Update dependencies?")) {
                 try {
                     const newPackageJson = packageJson;
 
@@ -221,12 +233,34 @@ async function checkMigration() {
                         }
                     }
                     // Writes the new package keeping the current format
-                    await writeJson(packageJsonPath, newPackageJson, { spaces: 2 });
+                    await writeJson(paths.packageJson, newPackageJson, { spaces: 2 });
+
+                    // Auto install is only implemented for npm right now
+                    if (!existsSync(paths.packageLock)) {
+                        console.log(
+                            "Dependency versions have been updated in the package.json of the widget.\n" +
+                                bold(
+                                    whiteBright(
+                                        "Force install dependencies with your package manager to complete the update."
+                                    )
+                                )
+                        );
+                        return;
+                    }
+
                     console.log("Deleting old dependencies...");
-                    execSync("shx rm -rf ./{node_modules,package-lock.json}", { cwd: process.cwd(), stdio: "inherit" });
-                    console.log("Done.");
+                    await rm(paths.node_modules, { recursive: true, force: true });
+                    await rm(paths.packageLock);
+
+                    console.log("Installing dependencies...");
                     execSync(`npm install`, { cwd: process.cwd(), stdio: "inherit" });
-                    execSync(`npx eslint --no-eslintrc --rule '@typescript-eslint/no-unused-vars: ["error", { enableAutofixRemoval: { imports: true } }]' --ext ts,tsx,js,jsx --parser @typescript-eslint/parser --plugin '@typescript-eslint' --fix ./src`, { cwd: process.cwd(), stdio: "inherit" })
+                    execSync(
+                        `npx eslint --no-config-lookup --rule '@typescript-eslint/no-unused-vars: ["error", { enableAutofixRemoval: { imports: true } }]' --ext ts,tsx,js,jsx --parser @typescript-eslint/parser --plugin '@typescript-eslint' --fix ./src`,
+                        { cwd: process.cwd(), stdio: "inherit" }
+                    );
+                    await auditPluggableWidgetsTools();
+
+                    console.log(green("Done auto-updating dependencies"));
                 } catch (e) {
                     console.log(red("An error occurred while auto updating your dependencies"));
                     console.error(e);